requirements.txt
boto3==1.34.0
firebase_admin==6.3.0
folium==0.16.0
matplotlib==3.7.1
numpy==1.24.3
opencv_contrib_python==4.7.0.72
opencv_python==4.8.0.76
opencv_python_headless==4.8.0.76
pandas==1.5.3
pandasai==2.0.40
prophet==1.1.5
protobuf==5.26.1
python-dotenv==1.0.1
scikit_learn==1.2.2
streamlit==1.31.1
streamlit_folium==0.19.0
streamlit_option_menu==0.3.12
streamlit_shadcn_ui==0.1.18
streamlit_survey==0.1.0
